Better color shadows, card height no longer enforced, card logic better
Card heights are no longer restricted to a max height of 120 px (that was dumb now that I think about it)

Code didn't always remove the "download" button or whatever for each card because it wasn't always the 2nd 'a' tag in the card. Code will not traverse the DOM more specifically to remove the correct element.